+++ KeepingYouAwake/KYAAppController.m
@@ -10,11 +10,16 @@
 #import "KYASleepWakeTimer.h"
 #import "KYAEventHandler.h"
 #import "KYAMenuBarIcon.h"
-#import "NSApplication+LoginItem.h"
+#import "KYABatteryStatus.h"
+#import "KYABatteryCapacityThreshold.h"
 #import "NSUserDefaults+Keys.h"
 
 @interface KYAAppController () <NSUserNotificationCenterDelegate>
-@property (strong, nonatomic, readwrite) KYASleepWakeTimer *sleepWakeTimer;
+@property (nonatomic, readwrite) KYASleepWakeTimer *sleepWakeTimer;
+
+// Battery Status
+@property (nonatomic) KYABatteryStatus *batteryStatus;
+@property (nonatomic, getter=isBatteryOverrideEnabled) BOOL batteryOverrideEnabled;
 
 // Status Item
 @property (strong, nonatomic) NSStatusItem *statusItem;
@@ -22,7 +27,6 @@
 // Menu
 @property (weak, nonatomic) IBOutlet NSMenu *menu;
 @property (weak, nonatomic) IBOutlet NSMenu *timerMenu;
-@property (weak, nonatomic) IBOutlet NSMenuItem *startAtLoginMenuItem;
 @end
 
 @implementation KYAAppController
@@ -51,6 +55,7 @@
                                                    object:nil
          ];
         
+        [self configureBatteryStatus];
         [self configureEventHandler];
     }
     return self;
@@ -58,22 +63,14 @@
 
 - (void)dealloc
 {
-    [[NSNotificationCenter defaultCenter] removeObserver:self name:NSApplicationDidFinishLaunchingNotification object:nil];
+    NSNotificationCenter *notificationCenter = [NSNotificationCenter defaultCenter];
+    [notificationCenter removeObserver:self name:NSApplicationDidFinishLaunchingNotification object:nil];
+    [notificationCenter removeObserver:self name:kKYABatteryCapacityThresholdDidChangeNotification object:nil];
 }
 
 - (void)awakeFromNib
 {
     [super awakeFromNib];
-    
-    // Check start at login state
-    if([[NSApplication sharedApplication] kya_isStartingAtLogin])
-    {
-        self.startAtLoginMenuItem.state = NSOnState;
-    }
-    else
-    {
-        self.startAtLoginMenuItem.state = NSOffState;
-    }
     
     [NSUserNotificationCenter defaultUserNotificationCenter].delegate = self;
 }
@@ -140,23 +137,6 @@
     [defaults synchronize];
 }
 
-#pragma mark - Start at Login
-
-- (IBAction)toggleStartAtLogin:(id)sender
-{
-    BOOL shouldStartAtLogin = ![[NSApplication sharedApplication] kya_isStartingAtLogin];
-    [NSApplication sharedApplication].kya_startAtLogin = shouldStartAtLogin;
-    
-    if(shouldStartAtLogin)
-    {
-        self.startAtLoginMenuItem.state = NSOnState;
-    }
-    else
-    {
-        self.startAtLoginMenuItem.state = NSOffState;
-    }
-}
-
 #pragma mark - Toggle Handling
 
 - (void)toggleStatus:(id)sender
@@ -186,12 +166,12 @@
     if(active)
     {
         button.image = menubarIcon.activeIcon;
-        button.toolTip = NSLocalizedString(@"Click to allow sleep\nRight click to show menu", nil);
+        button.toolTip = NSLocalizedString(@"Click to allow sleep\nRight click to show menu", @"Click to allow sleep\nRight click to show menu");
     }
     else
     {
         button.image = menubarIcon.inactiveIcon;
-        button.toolTip = NSLocalizedString(@"Click to prevent sleep\nRight click to show menu", nil);
+        button.toolTip = NSLocalizedString(@"Click to prevent sleep\nRight click to show menu", @"Click to prevent sleep\nRight click to show menu");
     }
 }
 
@@ -210,12 +190,19 @@
         return;
     }
     
+    // Check battery overrides and register for capacity changes.
+    [self checkAndEnableBatteryOverride];
+    if([[NSUserDefaults standardUserDefaults] kya_isBatteryCapacityThresholdEnabled])
+    {
+        [self.batteryStatus registerForCapacityChangesIfNeeded];
+    }
+    
     [self.sleepWakeTimer scheduleWithTimeInterval:timeInterval completion:^(BOOL cancelled) {
         // Post notifications
         if([[NSUserDefaults standardUserDefaults] kya_areNotificationsEnabled])
         {
             NSUserNotification *n = [NSUserNotification new];
-            n.informativeText = NSLocalizedString(@"Allowing your Mac to go to sleep…", nil);
+            n.informativeText = NSLocalizedString(@"Allowing your Mac to go to sleep…", @"Allowing your Mac to go to sleep…");
             [[NSUserNotificationCenter defaultUserNotificationCenter] scheduleNotification:n];
         }
     }];
@@ -227,7 +214,7 @@
         
         if(timeInterval == KYASleepWakeTimeIntervalIndefinite)
         {
-            n.informativeText = NSLocalizedString(@"Preventing your Mac from going to sleep…", nil);
+            n.informativeText = NSLocalizedString(@"Preventing your Mac from going to sleep…", @"Preventing your Mac from going to sleep…");
         }
         else
         {
@@ -236,7 +223,7 @@
             NSString *remainingTimeString = [formatter stringFromDate:[NSDate date] toDate:self.sleepWakeTimer.fireDate];
             formatter.includesTimeRemainingPhrase = YES;
             
-            n.informativeText = [NSString stringWithFormat:NSLocalizedString(@"Preventing your Mac from going to sleep for\n%@…", nil), remainingTimeString];
+            n.informativeText = [NSString stringWithFormat:NSLocalizedString(@"Preventing your Mac from going to sleep for\n%@…", @"Preventing your Mac from going to sleep for\n%@…"), remainingTimeString];
         }
         
         [[NSUserNotificationCenter defaultUserNotificationCenter] scheduleNotification:n];
@@ -245,17 +232,20 @@
 
 - (void)terminateTimer
 {
-    [self.sleepWakeTimer invalidate];
+    [self disableBatteryOverride];
+    [self.batteryStatus unregisterFromCapacityChanges];
+    
+    if([self.sleepWakeTimer isScheduled])
+    {
+        [self.sleepWakeTimer invalidate];
+    }
 }
 
 #pragma mark - Menu Delegate
 
 - (IBAction)selectTimeInterval:(NSMenuItem *)sender
 {
-    if([self.sleepWakeTimer isScheduled])
-    {
-        [self.sleepWakeTimer invalidate];
-    }
+    [self terminateTimer];
 	
 	if (sender.alternate)
 	{
@@ -336,6 +326,76 @@
     return YES;
 }
 
+#pragma mark - Battery Status
+
+- (KYABatteryStatus *)batteryStatus
+{
+    if(_batteryStatus == nil)
+    {
+        _batteryStatus = [KYABatteryStatus new];
+        
+        __weak typeof(self) weakSelf = self;
+        _batteryStatus.capacityChangeHandler = ^(CGFloat capacity) {
+            [weakSelf batteryCapacityDidChange:capacity];
+        };
+    }
+    return _batteryStatus;
+}
+
+- (void)configureBatteryStatus
+{
+    if(![self.batteryStatus isBatteryStatusAvailable])
+    {
+        return;
+    }
+    
+    [[NSNotificationCenter defaultCenter] addObserver:self
+                                             selector:@selector(batteryCapacityThresholdDidChange:)
+                                                 name:kKYABatteryCapacityThresholdDidChangeNotification
+                                               object:nil
+     ];
+    
+    // Start receiving battery status changes
+    if([[NSUserDefaults standardUserDefaults] kya_isBatteryCapacityThresholdEnabled])
+    {
+        [self.batteryStatus registerForCapacityChangesIfNeeded];
+    }
+}
+
+- (void)checkAndEnableBatteryOverride
+{
+    CGFloat currentCapacity = self.batteryStatus.currentCapacity;
+    CGFloat threshold = [NSUserDefaults standardUserDefaults].kya_batteryCapacityThreshold;
+    
+    self.batteryOverrideEnabled = (currentCapacity <= threshold);
+}
+
+- (void)disableBatteryOverride
+{
+    self.batteryOverrideEnabled = NO;
+}
+
+- (void)batteryCapacityDidChange:(CGFloat)capacity
+{
+    CGFloat threshold = [NSUserDefaults standardUserDefaults].kya_batteryCapacityThreshold;
+    if([self.sleepWakeTimer isScheduled] && (capacity <= threshold) && ![self isBatteryOverrideEnabled])
+    {
+        [self terminateTimer];
+    }
+}
+
+#pragma mark - Battery Capacity Threshold Changes
+
+- (void)batteryCapacityThresholdDidChange:(NSNotification *)notification
+{
+    if(![self.batteryStatus isBatteryStatusAvailable])
+    {
+        return;
+    }
+    
+    [self terminateTimer];
+}
+
 #pragma mark - Apple Event Manager
 
 - (void)applicationWillFinishLaunching:(NSNotification *)notification
@@ -346,7 +406,6 @@
                                                         andEventID:kAEGetURL
      ];
 }
-
 
 - (void)handleGetURLEvent:(NSAppleEventDescriptor *)event withReplyEvent:(NSAppleEventDescriptor *)reply
 {
@@ -359,32 +418,31 @@
 {
     __weak typeof(self) weakSelf = self;
     [[KYAEventHandler mainHandler] registerActionNamed:@"activate" block:^(KYAEvent *event) {
+        typeof(self) strongSelf = weakSelf;
+        
         NSDictionary *parameters = event.arguments;
         NSString *seconds = parameters[@"seconds"];
         NSString *minutes = parameters[@"minutes"];
         NSString *hours = parameters[@"hours"];
         
-        if([self.sleepWakeTimer isScheduled])
-        {
-            [self.sleepWakeTimer invalidate];
-        }
+        [self terminateTimer];
         
         // Activate indefinitely if there are no parameters
         if(parameters.count == 0)
         {
-            [weakSelf activateTimer];
+            [strongSelf activateTimer];
         }
         else if(seconds)
         {
-            [weakSelf activateTimerWithTimeInterval:(NSTimeInterval)ceil(seconds.doubleValue)];
+            [strongSelf activateTimerWithTimeInterval:(NSTimeInterval)ceil(seconds.doubleValue)];
         }
         else if(minutes)
         {
-            [weakSelf activateTimerWithTimeInterval:(NSTimeInterval)KYA_MINUTES(ceil(minutes.doubleValue))];
+            [strongSelf activateTimerWithTimeInterval:(NSTimeInterval)KYA_MINUTES(ceil(minutes.doubleValue))];
         }
         else if(hours)
         {
-            [weakSelf activateTimerWithTimeInterval:(NSTimeInterval)KYA_HOURS(ceil(hours.doubleValue))];
+            [strongSelf activateTimerWithTimeInterval:(NSTimeInterval)KYA_HOURS(ceil(hours.doubleValue))];
         }
     }];
     [[KYAEventHandler mainHandler] registerActionNamed:@"deactivate" block:^(KYAEvent *event) {