+++ KeepingYouAwake/KYAGeneralPreferencesViewController.h
@@ -1,0 +1,12 @@
+//
+//  KYAGeneralPreferencesViewController.h
+//  KeepingYouAwake
+//
+//  Created by Marcel Dierkes on 18.12.15.
+//  Copyright © 2015 Marcel Dierkes. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+@interface KYAGeneralPreferencesViewController : NSViewController
+@end